Add signout testing functionality and create sub-node at login with Google account.
Added signout testing functionality and created a sub-node at login with Google account, including user information.

The signout feature has been implemented and can be accessed by clicking the button in the upper right corner with the profile photo of the logged-in user (CRITICAL CHANGE IN TESTING PHASE).

Upon initial login, the system creates a user with a unique identifier (UID) and generates a sub-node containing the user's name and email address (CRITICAL CHANGE IN TESTING PHASE).